reshuffled some stuff, corrected a capital letter
